call unscheduleUpdate() in physics::Sprite::scheduleUpdate() first
Wait? What? This is crazy! Right. It is. But it's common to inherit
from physics::Sprite and, if you want to do something in update(), to call
scheduleUpdate() in your init() method. This is the cocos2dx-way to do it.

But (!!) the internal scheduler will crash with a not very clear error
message, if you call scheduleUpdate() twice! And this is exactly what I want
to prevent. And yes, scheduleUpdate() is not virtual. That's another bug -.-