Bugfix: Random crahses when the CCBLoader is used.
We need to set an owner to the CCBLoader. In readNodeGraphFromFile we set the our CCBLoader Object as owner - the CCBReader puts this object into the autorelease manager. After finishing the loading we destroy ccbloader cause it’s no longer required, but the AutoreleasePool already registered the object and tries to destroy it again at a later point which causes an memory access violation.