Bugfix: This is dirty at the moment, but until now I hadn't found a better solution.
The problem was that when I leave a scene cocos cleans up all childs of this scene. For example the Box2dContainer. This will destroy the attached b2World which will destroy all b2Bodies. Later when the sprites with the bodies attached are destroyed again the app crashes. We cannot use the reference counter/cocos autorelease pool because box2d has it's own memory managment. We cannot use smart pointers because it will not work together well neither. My solution for now is to remove all children from the Box2dContainer when it is removed from the scene (yes, this can cause ugly side effects). I assume that all physic sprites are childs of the box2dcontainer.

We really need a better solution for this.